weird_number_generator -> weird_number_finder
- moved print_weird_numbers method from weird_number_generator to weird_numbers because it just prints them, and rename the generator file.
- The main logic to be tested is in weird_number_finder, and that file is now smaller.
