+++ weird_numbers.py
@@ -1,5 +1,5 @@
 import time
-from weird_number_generator import print_weird_numbers
+from weird_number_finder import is_weird
 
 
 def main():
@@ -16,5 +16,17 @@ def main():
         print("Found the first ", n, " weird numbers in ", elapsed, "seconds")
     except ValueError:
         print("That was not a valid integer.")
+
+
+def print_weird_numbers(n):
+    candidate = 2
+    while n > 0:
+        if is_weird(candidate) == 1:
+            end = "\n" if n % 20 == 0 else " "
+            print(candidate, end=end)
+            n = n - 1
+        candidate = candidate + 1
+    print("\ndone.")
+
 
 main()
